move admin tool dashboard to its own chunk
MIMIR-475
