+++ react4xp.properties
@@ -68,6 +68,6 @@ overwriteConstantsFile = true
 #       XP controllers etc, which can be .js or .es6. Default value if not changed is jsx,js,tsx,ts,es6,es. Also note that
 #       tsx/ts files are NOT supported out of the box. Rules for typescript compilation must be added in your own
 #       overrideComponentWebpack file (see above).
-chunkDirs =
+chunkDirs = dashboard
 entryDirs = _entries
 entryExtensions =
