+++ src/main/resources/react4xp/dashboard/components/ConnectionBadge.jsx
