+++ src/main/resources/react4xp/dashboard/components/DataQueryBadges.jsx
