+++ src/main/resources/react4xp/dashboard/components/ReactTable.jsx
