+++ src/main/resources/react4xp/dashboard/components/RefreshStatisticsForm.jsx
