+++ src/main/resources/react4xp/dashboard/components/RefreshStatisticsModal.jsx
