+++ src/main/resources/react4xp/dashboard/components/RefreshStatisticsStatus.jsx
