+++ src/main/resources/react4xp/dashboard/containers/DataSources/actions.es6
