+++ src/main/resources/react4xp/dashboard/containers/DataSources/slice.es6
