+++ src/main/resources/react4xp/dashboard/containers/HomePage/actions.es6
