+++ src/main/resources/react4xp/dashboard/containers/HomePage/selectors.es6
