+++ src/main/resources/react4xp/dashboard/containers/HomePage/slice.es6
