+++ src/main/resources/react4xp/dashboard/containers/Jobs/actions.es6
