+++ src/main/resources/react4xp/dashboard/containers/Jobs/selectors.es6
