+++ src/main/resources/react4xp/dashboard/containers/Jobs/slice.es6
