+++ src/main/resources/react4xp/dashboard/containers/StatRegDashboard/actions.es6
