+++ src/main/resources/react4xp/dashboard/containers/StatRegDashboard/index.jsx
