+++ src/main/resources/react4xp/dashboard/containers/StatRegDashboard/listeners.es6
