+++ src/main/resources/react4xp/dashboard/containers/StatRegDashboard/selectors.es6
