+++ src/main/resources/react4xp/dashboard/containers/StatRegDashboard/slice.es6
