+++ src/main/resources/react4xp/dashboard/containers/Statistics/StatisticsLog.jsx
