+++ src/main/resources/react4xp/dashboard/containers/Statistics/actions.es6
