+++ src/main/resources/react4xp/dashboard/containers/Statistics/index.jsx
