+++ src/main/resources/react4xp/dashboard/containers/Statistics/listeners.es6
