+++ src/main/resources/react4xp/dashboard/containers/Statistics/selectors.es6
