+++ src/main/resources/react4xp/dashboard/containers/Statistics/slice.es6
