+++ src/main/resources/react4xp/dashboard/store/configureStore.es6
