+++ src/main/resources/react4xp/dashboard/store/reducers.es6
