+++ src/main/resources/react4xp/dashboard/utils/websocket/WebsocketProvider.jsx
